+++ Source/EventFlow/ReadStores/ReadModelPopulator.cs
@@ -23,7 +23,6 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Data;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
@@ -48,7 +47,7 @@ namespace EventFlow.ReadStores
         private readonly IServiceProvider _serviceProvider;
         private readonly IEventUpgradeContextFactory _eventUpgradeContextFactory;
         private readonly IMemoryCache _memoryCache;
-        private ConcurrentQueue<AllEventsPage> _pipedEvents = new ConcurrentQueue<AllEventsPage>();
+        private readonly ConcurrentQueue<AllEventsPage> _pipedEvents = new ConcurrentQueue<AllEventsPage>();
 
         public ReadModelPopulator(
             ILogger<ReadModelPopulator> logger,
@@ -118,14 +117,15 @@ namespace EventFlow.ReadStores
             var combinedReadModelTypeString = string.Join(", ", readModelTypes.Select(type => type.PrettyPrint()));
             _logger.LogInformation("Starting populating of {ReadModelTypes}", combinedReadModelTypeString);
 
-            var loadEventsTasks = LoadEvents(cancellationToken);
-            var processEventQueueTask = ProcessEventQueue(readModelTypes, cancellationToken);
+            var loadEventsTasks = LoadEventsAsync(cancellationToken);
+            var processEventQueueTask = ProcessEventQueueAsync(readModelTypes, cancellationToken);
+
             await Task.WhenAll(loadEventsTasks, processEventQueueTask);
 
-            _logger.LogInformation("Population of readmodels completed");
+            _logger.LogInformation("Population of read models completed");
         }
 
-        private async Task LoadEvents(CancellationToken cancellationToken)
+        private async Task LoadEventsAsync(CancellationToken cancellationToken)
         {
             long totalEvents = 0;
             var currentPosition = GlobalPosition.Start;
@@ -162,10 +162,11 @@ namespace EventFlow.ReadStores
             }
         }
 
-        private async Task ProcessEventQueue(IReadOnlyCollection<Type> readModelTypes, CancellationToken cancellationToken)
+        private async Task ProcessEventQueueAsync(
+            IReadOnlyCollection<Type> readModelTypes,
+            CancellationToken cancellationToken)
         {
             var domainEventsToProcess = new List<IDomainEvent>();
-            AllEventsPage fetchedEvents;
 
             var hasMoreEvents = true;
             do
@@ -173,11 +174,11 @@ namespace EventFlow.ReadStores
                 var noEventsToReady = !_pipedEvents.Any();
                 if (noEventsToReady)
                 {
-                    await Task.Delay(100);
+                    await Task.Delay(100, cancellationToken);
                     continue;
                 }
 
-                _pipedEvents.TryDequeue(out fetchedEvents);
+                _pipedEvents.TryDequeue(out var fetchedEvents);
                 if (fetchedEvents == null)
                 {
                     continue;
@@ -190,7 +191,7 @@ namespace EventFlow.ReadStores
                 var processEvents = !hasMoreEvents || batchExceedsThreshold;
                 if (processEvents)
                 {
-                    var readModelUpdateTasks = readModelTypes.Select(readModelType => ProcessEvents(readModelType, domainEventsToProcess, cancellationToken));
+                    var readModelUpdateTasks = readModelTypes.Select(readModelType => ProcessEventsAsync(readModelType, domainEventsToProcess, cancellationToken));
                     await Task.WhenAll(readModelUpdateTasks);
                     
                     domainEventsToProcess.Clear();
@@ -199,7 +200,10 @@ namespace EventFlow.ReadStores
             while (hasMoreEvents);
         }
 
-        private async Task ProcessEvents(Type readModelType, IReadOnlyCollection<IDomainEvent> processEvents, CancellationToken cancellationToken)
+        private async Task ProcessEventsAsync(
+            Type readModelType,
+            IReadOnlyCollection<IDomainEvent> processEvents,
+            CancellationToken cancellationToken)
         {
             try
             {
@@ -209,11 +213,12 @@ namespace EventFlow.ReadStores
 
                 var readModelTypes = new[]
                 {
-                    typeof( IAmReadModelFor<,,> )
+                    typeof(IAmReadModelFor<,,> )
                 };
 
-                var aggregateEventTypes = _memoryCache.GetOrCreate(CacheKey.With(GetType(), readModelType.ToString(), nameof(ProcessEvents)), 
-                    e => new HashSet<Type>(readModelType.GetTypeInfo()
+                var aggregateEventTypes = _memoryCache.GetOrCreate(
+                    CacheKey.With(GetType(), readModelType.ToString(), nameof(ProcessEventsAsync)), 
+                    _ => new HashSet<Type>(readModelType.GetTypeInfo()
                         .GetInterfaces()
                         .Where(i => i.GetTypeInfo().IsGenericType && readModelTypes.Contains(i.GetGenericTypeDefinition()))
                         .Select(i => i.GetTypeInfo().GetGenericArguments()[2])));
@@ -249,7 +254,7 @@ namespace EventFlow.ReadStores
             }
             catch (Exception e)
             {
-                _logger.LogWarning($"Exception when populating: {readModelType}. Details: {e}");
+                _logger.LogError(e, $"Exception when populating: {readModelType}");
             }
         }